cambios en los microservicios y api gateway terminado

+++ Backend/api-gateway/src/typeDefs/product_type_defs.js
@@ -9,7 +9,7 @@ const productTypeDefs = gql `
         stock: Int!
         description: String!
         category: String!
-        sales: Float!
+        sales: Int!
     }
 
     type paginationProducts {
